requirements.txt
django
djangorestframework
djangorestframework-jwt
cryptography
pyjwt
python-dotenv
python-jose
gunicorn
django-money